+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/createaccountingcreditnote.ts
@@ -58,5 +58,5 @@ export class CreateAccountingCreditNoteResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }