+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/createconnection.ts
@@ -46,5 +46,5 @@ export class CreateConnectionResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }