+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/getconfigtextsyncflow.ts
@@ -35,5 +35,5 @@ export class GetConfigTextSyncFlowResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }