+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/getsyncstatus.ts
@@ -37,7 +37,7 @@ export class GetSyncStatusResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Success