+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/listcommerceorders.ts
@@ -73,5 +73,5 @@ export class ListCommerceOrdersResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }