+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/listcommercetransactions.ts
@@ -73,5 +73,5 @@ export class ListCommerceTransactionsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }