+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/refreshcompanyinfo.ts
@@ -43,5 +43,5 @@ export class RefreshCompanyInfoResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }