+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/updateconfigtextsyncflow.ts
@@ -35,5 +35,5 @@ export class UpdateConfigTextSyncFlowResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }