+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/updateconnection.ts
@@ -52,5 +52,5 @@ export class UpdateConnectionResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }