+++ previous-versions/sync-for-commerce-version-1/src/sdk/sdk.ts
@@ -78,9 +78,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.1";
-    sdkVersion = "0.5.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 0.5.0 2.188.1 1.1 @codat/sync-for-commerce-version-1";
+    sdkVersion = "0.5.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 0.5.1 2.194.1 1.1 @codat/sync-for-commerce-version-1";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);