+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [typescript v0.5.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [NPM v0.5.0] https://www.npmjs.com/package/@codat/sync-for-commerce-version-1/v/0.5.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-11-20 01:28:50
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.1] previous-versions/sync-for-commerce-version-1
+### Releases
+- [NPM v0.5.1] https://www.npmjs.com/package/@codat/sync-for-commerce-version-1/v/0.5.1 - previous-versions/sync-for-commerce-version-1